add some feature

+++ Project/src/Home.js
@@ -13,8 +13,11 @@ async function addProductStructureToHtml(productList) {
 
         productElement.innerHTML = `
             <ul id="ProdProperty">
-                <li class="prodAttribute">Tên sản phẩm : ${productsName}</li>
-                <li class="prodAttribute">Giá tiền : ${currentPrice}</li>
+                <li class="prodAttribute">
+                Tên sản phẩm : ${productsName}
+                </li>
+                <li class="prodAttribute">Giá tiền : ${currentPrice} đ
+                </li>
                 <li class="prodAttribute">Ngày tạo : ${dateCreate}</li>
                 <i class="fa-solid fa-trash-can deleteButton" id="${id}"></i>
             </ul>
@@ -22,23 +25,29 @@ async function addProductStructureToHtml(productList) {
 
         const deleteButton = productElement.querySelector(".deleteButton");
         deleteButton.addEventListener("click", async (event) => {
-            const prodID = event.target.id;
-            try {
-                const response = await fetch(`http://localhost:8080/api/product/${prodID}`, {
-                    method: "DELETE",
-                    headers: {
-                        "Content-Type": "application/json"
-                    },
-                });
-
-                if (!response.ok) {
-                    throw new Error("Network response was not ok");
-                } else {
-                    console.log("Xóa thành công");
+            if(confirm("Bạn có muốn xóa không")){
+                const prodID = event.target.id;
+                try {
+                    const response = await fetch(`http://localhost:8080/api/product/${prodID}`, {
+                        method: "DELETE",
+                        headers: {
+                            "Content-Type": "application/json"
+                        },
+                    });
+    
+                    if (!response.ok) {
+                        throw new Error("Network response was not ok");
+                    } else {
+                        if(confirm("Đã xóa thành công bạn có muốn load lại trang để thấy sự thay đổi hay không")) {
+                            location.reload();
+                        }
+                        
+                    }
+                } catch (error) {
+                    console.error('Error:', error);
                 }
-            } catch (error) {
-                console.error('Error:', error);
             }
+          
         });
 
         wrapProd.appendChild(productElement);
